Day 10, Part 02
Part 02 had to be run twice - with 'flip' True and False, because
I couldn't figure out how to calculate which side of the loop is
the 'inside'.

I suppose I could refactor the code to run twice and return
the figure out which solution is 'right', either using a naive
approach (which is smaller?) or by checking the results to see
if they are fully contained within the loop.

But you know what? The goal is to get the right answer, and I did.

Ship it.