Day 12, Part 2 redo
I don't have a CS degree - I never heard of dynamic programming until
this week. Although it seems to be recursion's cousin.

So I couldn't wrap my head around the examples that did dynamic
programming until I saw this one on Reddit by /4HbQ

https://www.reddit.com/r/adventofcode/comments/18ge41g/comment/kd0ki5t/

It is elegant - eschewing the atomic incrementing I saw in other
solutions in favor of advancing a block at a time by checking if the
block can fit. Even with a cache, it takes 0.5 seconds.

I don't know if I could have come up with a solution like this on my
own, at least not without spending a week on the problem. But I have
work, sleep, and other puzzles to solve.