Day 21, Part 21 revisited
I could not figure out how to solve this one before the heat death of
the universe, so I turned to reddit. My solution is based on the one
by /Thomasjevskij since his was the one I could grok.

I doubt I would have made the leap to "oh, it's quadratic!" on my own,
and I barely remember high school algebra, let alone high school.

https://www.reddit.com/r/adventofcode/comments/18nevo3/comment/ked8epo/?rdt=41987

Interesting tactic with the alternating queues - is that a C-style
optimization?