Day 24, Part 02, wip
This solution will run in approximately O(Heat Death of the Universe).

Given three arbitrary points, for the six permutations of hit order,
search for when the points get hit, for times x, y, z, where x < y < z.

The problem with this algorithm is that the search space is exponential
- 10 ticks means at most 6000 iterations, 100 ticks means at most 6 million iterations. If the rock hit every hailstone one after another, it would take a minimum of 301 ticks.

What I need to do is evaluate the points and figure out which three
points are the "best" to evaluate. I don't know how to do that, though.